Create new stream for plugins
Previously the all repositories were using the same plugins stream
object.
It caused that the modules were mixed from diferrent sources and that
caused that when calculated hash for a module and its dependencies, it
may found a module from the other repository, and the hash may not
updated.
Issue was fixed by creating a new plugin stream for each repository.